requirements.txt
Flask==2.0.2
gunicorn==20.1.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.26
psycopg2-binary==2.9.1
python-decouple==3.5
flask-blueprint==1.3.0
Flask-Migrate==3.1.0
ariadne==0.14.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0